+++ GMLSCG/GMLS.h
@@ -58,7 +58,7 @@ struct All_Values
 struct Coalition
 {
 	int index =-1;
-	int cost = -1;
+	double cost = -1;
 	vector<int> pattern;
 };
 
@@ -67,9 +67,9 @@ struct All_Lists
 	vector<Parameter_Struct> primal_parameters; // 物品时间角标
 	vector<Variabe_Struct> current_solns_list; // 初始求解结果
 
-	vector<int> MP_solns_list;
+	vector<double> MP_solns_list;
 
-	vector<float>dual_prices_list;
+	vector<double>dual_prices_list;
 
 	vector<vector<int>> X_fix;
 	vector<vector<int>> Y_fix;
@@ -79,13 +79,13 @@ struct All_Lists
 
 	vector<int> single_machine_demand;
 
-	vector<int> coalition_cost_list; // list of all possible c(S)
+	vector<double> coalition_cost_list; // list of all possible c(S)
 	
 	vector<vector<int>> pattern_list;
 	vector<int> pattern_idx_list;
 	vector<int> pattern_cost_list;
 
-	vector<int> SP_solns_list; // results of master problem
+	vector<double> SP_solns_list; // results of master problem
 
 	vector<vector<int>> model_matrix; // master problem matrix modeled by rows 
 	vector<vector<int>> Master_Matrix_C; // master problem matrix modeled by cols
