+++ GMLSCG/OriginalProblem.cpp
@@ -118,28 +118,28 @@ void SolveOriginalProblem(All_Values& Values, All_Lists& Lists, int coalition_fl
 
 		printf("\n/////////// CPLEX SOLVING END ////////////\n");
 
-		int Obj_value = Cplex_OP.getObjValue();
+		double Obj_value = Cplex_OP.getObjValue();
 		Lists.coalition_cost_list.push_back(Obj_value);
 		Lists.coalitions_list[coalition_flag].cost = Obj_value;
 
-		printf("\n\t Obj = %d\n", Obj_value);
+		printf("\n\t Obj = %f\n", Obj_value);
 		cout << endl;
 
 		for (int t = 0; t < T_num; t++) {
-			int soln_val = Cplex_OP.getValue(X_vars_list[t]);
-			printf("\t X_%d = %d\n", t + 1, soln_val);
+			double soln_val = Cplex_OP.getValue(X_vars_list[t]);
+			printf("\t X_%d = %f\n", t + 1, soln_val);
 		}
 		cout << endl;
 
 		for (int t = 0; t < T_num; t++) {
-			int soln_val = Cplex_OP.getValue(I_vars_list[t]);
-			printf("\t I_%d= %d\n", t + 1, soln_val);
+			double soln_val = Cplex_OP.getValue(I_vars_list[t]);
+			printf("\t I_%d= %f\n", t + 1, soln_val);
 		}
 		cout << endl;
 
 		for (int t = 0; t < T_num; t++) {
-			int soln_val = Cplex_OP.getValue(Y_vars_list[t]);
-			printf("\t Y_%d= %d\n", t + 1, soln_val);
+			double soln_val = Cplex_OP.getValue(Y_vars_list[t]);
+			printf("\t Y_%d= %f\n", t + 1, soln_val);
 		}
 	}
