+++ GMLSRG/FirstMasterProblem.cpp
@@ -99,8 +99,8 @@ void SolveFirstMasterProblem(All_Values& Values, All_Lists& Lists) {
 		printf("\n\t /////////// W /////////\n\n");
 
 		for (int col = 0; col < all_cols_num; col++) {
-			int w_soln_val = Cplex_MP.getValue(W_Vars[col]);
-			printf("\t W_%d = %d\n", col + 1, w_soln_val);
+			double w_soln_val = Cplex_MP.getValue(W_Vars[col]);
+			printf("\t W_%d = %f\n", col + 1, w_soln_val);
 			Lists.MP_solns_list.push_back(w_soln_val);
 		}
 	}
